[IMP] Canonical enumeration type.
