+++ components-pro/code-area/enum.tsx
@@ -1,0 +1,4 @@
+export enum ThemeSwitch {
+  idea = 'idea',
+  material = 'material',
+}
