+++ components-pro/performance-table/enum.tsx
@@ -1,0 +1,4 @@
+export enum TableQueryBarType {
+  professionalBar = 'professionalBar',
+  filterBar = 'filterBar',
+}
