+++ components-pro/trigger-field/enum.tsx
@@ -1,0 +1,5 @@
+export enum TriggerViewMode {
+  popup = 'popup',
+  drawer = 'drawer',
+  modal = 'modal',
+}
