+++ components/bar-code/enum.tsx
@@ -1,0 +1,4 @@
+export enum CodeType {
+  QR = 'qr',
+  BAR = 'bar',
+}
