+++ components/cascader/enum.tsx
@@ -1,0 +1,4 @@
+export enum MenuMode {
+  single = 'single',
+  multiple = 'multiple',
+}
