+++ components/image-crop/enum.tsx
@@ -1,0 +1,4 @@
+export enum ShapeCroper {
+  rect = 'rect',
+  round = 'round',
+}
