+++ components/table/enum.tsx
@@ -1,0 +1,4 @@
+export enum TableAutoHeightType {
+  maxHeight = 'maxHeight',
+  minHeight = 'minHeight',
+}
